// src/_proto/Protos/novel_announcements.ts
export interface GetNovelAnnouncementsRequest {
    /**
     * @generated from protobuf oneof: novelSelector
     */
    novelSelector: {
        oneofKind: "novelId";
        /**
         * @generated from protobuf field: int32 novelId = 1;
         */
        novelId: number;
    } | {
        oneofKind: "novelSlug";
        /**
         * @generated from protobuf field: string novelSlug = 2;
         */
        novelSlug: string;
    } | {
        oneofKind: undefined;
    };
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.PageInfoRequest pageInfo = 3;
     */
    pageInfo?: PageInfoRequest;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.GetNovelAnnouncementsRequest.Filter filter = 4;
     */
    filter?: GetNovelAnnouncementsRequest_Filter;
}
/**
 * @generated from protobuf message wuxiaworld.api.v2.GetNovelAnnouncementsRequest.Filter
 */
export interface GetNovelAnnouncementsRequest_Filter {
    /**
     * @generated from protobuf field: google.protobuf.BoolValue isSponsorsOnly = 1;
     */
    isSponsorsOnly?: BoolValue;
}
// ...
class GetNovelAnnouncementsRequest$Type extends MessageType<GetNovelAnnouncementsRequest> {
    constructor() {
        super("wuxiaworld.api.v2.GetNovelAnnouncementsRequest", [
            { no: 1, name: "novelId", kind: "scalar", oneof: "novelSelector", T: 5 /*ScalarType.INT32*/ },
            { no: 2, name: "novelSlug", kind: "scalar", oneof: "novelSelector", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "pageInfo", kind: "message", T: () => PageInfoRequest, options: { "wuxiaworld.api.v2.required": { value: true } } },
            { no: 4, name: "filter", kind: "message", T: () => GetNovelAnnouncementsRequest_Filter }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.GetNovelAnnouncementsRequest
 */
export const GetNovelAnnouncementsRequest = new GetNovelAnnouncementsRequest$Type();
// @generated message type with reflection information, may provide speed optimized methods
class GetNovelAnnouncementsRequest_Filter$Type extends MessageType<GetNovelAnnouncementsRequest_Filter> {
    constructor() {
        super("wuxiaworld.api.v2.GetNovelAnnouncementsRequest.Filter", [
            { no: 1, name: "isSponsorsOnly", kind: "message", T: () => BoolValue }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.GetNovelAnnouncementsRequest.Filter
 */
export const GetNovelAnnouncementsRequest_Filter = new GetNovelAnnouncementsRequest_Filter$Type();